Add an explanation section
I'd like to see an Explanation section in each RFC.

The Explanation is a long-winded and more casual styled, explanation of
the various details and implications of the RFC.

In contrast, the Problem and Proposal sections, should be more short,
to the point and use a more legalese style.

...

This RFC, while missing the Explanation, had Problem and Proposal
sections that were fairly on point.

Other RFCs have Proposal sections that are big with no Explanation
sections. I'd prefer the Proposal to get immediately to the technical
change, and move the rest to the Explanation.