Change the Synopsis to a table format
This is an experiment to see if it looks better.

Also move the See Also references into the Synopsis if they are pointing
to other RFCs. I think it is important to have pointers to related and
required RFCs up top.

We can still have a See Also section if it points to external info.